+++ src/components/OfflineIndicator.js
